Further reduce Markov table size

+++ uwurandom_types.h
@@ -14,13 +14,13 @@ typedef uint32_t uwu_random_number;
 
 typedef struct uwu_markov_choice uwu_markov_choice;
 struct uwu_markov_choice {
-    uint16_t next_ngram;
-    uint16_t cumulative_probability;
+    uint8_t next_ngram;
+    uint8_t cumulative_probability;
 };
 
 typedef struct {
     uint16_t choices;
-    uint16_t total_probability;
+    uint8_t total_probability;
     char character;
 } uwu_markov_ngram;
